requirements.txt
mediapipe==0.10.33
opencv-contrib-python==4.13.0.92
numpy==2.4.4
Pillow==12.2.0

